fix: fixed bug: number of replies rerender when user post a message to thread
